requirements.txt
django
numpy
Pillow
tensorflow
opencv-python
imutils
djangorestframework
scikit-image
